fix : RankListItem 컴포넌트 삭제
